+++ src/NexusMods.App.UI/WorkspaceSystem/Page/IPageFactoryParameter.cs
@@ -1,0 +1,6 @@
+namespace NexusMods.App.UI.WorkspaceSystem;
+
+/// <summary>
+/// Interface for parameters of <see cref="IPageFactory"/>
+/// </summary>
+public interface IPageFactoryParameter { }
